+++ lib/rubocop/cop/style/lambda.rb
@@ -22,11 +22,11 @@ module RuboCop
 
           return unless block_method == TARGET
           selector = block_method.loc.selector.source
-          lambda_length = lambda_length(node)
+          length = lambda_length(node)
 
-          if selector != '->' && lambda_length == 0
+          if selector != '->' && length == 1
             add_offense_for_single_line(block_method, args)
-          elsif selector == '->' && lambda_length > 0
+          elsif selector == '->' && length > 1
             add_offense(block_method, :expression, MULTI_MSG)
           end
         end
@@ -45,7 +45,7 @@ module RuboCop
           start_line = block_node.loc.begin.line
           end_line = block_node.loc.end.line
 
-          end_line - start_line
+          end_line - start_line + 1
         end
 
         def autocorrect(node)
@@ -62,11 +62,15 @@ module RuboCop
 
         def autocorrect_new_to_old(corrector, node)
           block_method, args = *node
+          # Avoid correcting to `lambdado` by inserting whitespace
+          # if none exists before or after the lambda arguments.
+          if needs_whitespace?(block_method, args, node)
+            corrector.insert_before(node.loc.begin, ' ')
+          end
           corrector.replace(block_method.loc.expression, 'lambda')
+          corrector.remove(args.loc.expression) if args.loc.expression
           return if args.children.empty?
-
           arg_str = " |#{lambda_arg_string(args)}|"
-          corrector.remove(args.loc.expression)
           corrector.insert_after(node.loc.begin, arg_str)
         end
 
@@ -79,6 +83,15 @@ module RuboCop
           whitespace_and_old_args = node.loc.begin.end.join(args.loc.end)
           corrector.insert_after(block_method.loc.expression, arg_str)
           corrector.remove(whitespace_and_old_args)
+        end
+
+        def needs_whitespace?(block_method, args, node)
+          selector_end = block_method.loc.selector.end.end_pos
+          args_begin   = args.loc.begin && args.loc.begin.begin_pos
+          args_end     = args.loc.end && args.loc.end.end_pos
+          block_begin  = node.loc.begin.begin_pos
+          (block_begin == args_end && selector_end == args_begin) ||
+            (block_begin == selector_end)
         end
 
         def lambda_arg_string(args)