+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 * Handle explicit `begin` blocks in `Lint/Void`. ([@bbatsov][])
 * Handle symbols in `Lint/Void`. ([@bbatsov][])
 * [#1695](https://github.com/bbatsov/rubocop/pull/1695): Fix bug with `--auto-gen-config` and `SpaceInsideBlockBraces`. ([@meganemura][])
+* Correct issues with whitespace around multi-line lambda arguments. ([@zvkemp][])
 
 ### Changes
 
@@ -1299,3 +1300,4 @@
 [@palkan]: https://github.com/palkan
 [@jdoconnor]: https://github.com/jdoconnor
 [@meganemura]: https://github.com/meganemura
+[@zvkemp]: https://github.com/zvkemp